+++ ext-src/stubs/php_swoole_thread_arginfo.h
@@ -1,5 +1,5 @@
 /* This is a generated file, edit the .stub.php file instead.
- * Stub hash: 864a92efe87d024e2084ae162912d64cf1498290 */
+ * Stub hash: 47148b0067d99c44ab50739f4184b021db929a6a */
 
 ZEND_BEGIN_ARG_INFO_EX(arginfo_class_Swoole_Thread___construct, 0, 0, 0)
 ZEND_END_ARG_INFO()
@@ -21,64 +21,3 @@ ZEND_END_ARG_INFO()
 
 ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_class_Swoole_Thread_getId, 0, 0, IS_LONG, 0)
 ZEND_END_ARG_INFO()
-
-#define arginfo_class_Swoole_Thread_Map___construct arginfo_class_Swoole_Thread___construct
-
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_class_Swoole_Thread_Map_offsetGet, 0, 1, IS_MIXED, 0)
-	ZEND_ARG_TYPE_INFO(0, key, IS_MIXED, 0)
-ZEND_END_ARG_INFO()
-
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_class_Swoole_Thread_Map_offsetExists, 0, 1, _IS_BOOL, 0)
-	ZEND_ARG_TYPE_INFO(0, key, IS_MIXED, 0)
-ZEND_END_ARG_INFO()
-
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_class_Swoole_Thread_Map_offsetSet, 0, 2, IS_VOID, 0)
-	ZEND_ARG_TYPE_INFO(0, key, IS_MIXED, 0)
-	ZEND_ARG_TYPE_INFO(0, value, IS_MIXED, 0)
-ZEND_END_ARG_INFO()
-
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_class_Swoole_Thread_Map_offsetUnset, 0, 1, IS_VOID, 0)
-	ZEND_ARG_TYPE_INFO(0, key, IS_MIXED, 0)
-ZEND_END_ARG_INFO()
-
-#define arginfo_class_Swoole_Thread_Map_count arginfo_class_Swoole_Thread_getId
-
-#define arginfo_class_Swoole_Thread_Map_keys arginfo_class_Swoole_Thread_getArguments
-
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_class_Swoole_Thread_Map_clean, 0, 0, IS_VOID, 0)
-ZEND_END_ARG_INFO()
-
-#define arginfo_class_Swoole_Thread_Map___wakeup arginfo_class_Swoole_Thread_Map_clean
-
-#define arginfo_class_Swoole_Thread_ArrayList___construct arginfo_class_Swoole_Thread___construct
-
-#define arginfo_class_Swoole_Thread_ArrayList_offsetGet arginfo_class_Swoole_Thread_Map_offsetGet
-
-#define arginfo_class_Swoole_Thread_ArrayList_offsetExists arginfo_class_Swoole_Thread_Map_offsetExists
-
-#define arginfo_class_Swoole_Thread_ArrayList_offsetSet arginfo_class_Swoole_Thread_Map_offsetSet
-
-#define arginfo_class_Swoole_Thread_ArrayList_offsetUnset arginfo_class_Swoole_Thread_Map_offsetUnset
-
-#define arginfo_class_Swoole_Thread_ArrayList_count arginfo_class_Swoole_Thread_getId
-
-#define arginfo_class_Swoole_Thread_ArrayList_clean arginfo_class_Swoole_Thread_Map_clean
-
-#define arginfo_class_Swoole_Thread_ArrayList___wakeup arginfo_class_Swoole_Thread_Map_clean
-
-#define arginfo_class_Swoole_Thread_Queue___construct arginfo_class_Swoole_Thread___construct
-
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_class_Swoole_Thread_Queue_push, 0, 1, IS_VOID, 0)
-	ZEND_ARG_TYPE_INFO(0, value, IS_MIXED, 0)
-	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, notify_which, IS_LONG, 0, "0")
-ZEND_END_ARG_INFO()
-
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_class_Swoole_Thread_Queue_pop, 0, 0, IS_MIXED, 0)
-	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, wait, IS_DOUBLE, 0, "0")
-ZEND_END_ARG_INFO()
-
-#define arginfo_class_Swoole_Thread_Queue_count arginfo_class_Swoole_Thread_getId
-
-#define arginfo_class_Swoole_Thread_Queue_clean arginfo_class_Swoole_Thread_Map_clean
-
-#define arginfo_class_Swoole_Thread_Queue___wakeup arginfo_class_Swoole_Thread_Map_clean
