+++ ext-src/stubs/php_swoole_thread_queue.stub.php
@@ -1,0 +1,11 @@
+<?php
+namespace Swoole\Thread {
+    class Queue implements Countable {
+        public function __construct() {}
+        public function push(mixed $value, int $notify_which = 0): void {}
+        public function pop(float $wait = 0): mixed {}
+        public function count(): int {}
+        public function clean(): void {}
+        public function __wakeup(): void {}
+    }
+}
