+++ ext-src/swoole_lock.cc
@@ -15,7 +15,6 @@
 */
 
 #include "php_swoole_private.h"
-#include "php_swoole_thread.h"
 #include "swoole_memory.h"
 #include "swoole_lock.h"
 
@@ -35,23 +34,8 @@ using swoole::RWLock;
 static zend_class_entry *swoole_lock_ce;
 static zend_object_handlers swoole_lock_handlers;
 
-#ifdef SW_THREAD
-struct LockResource: public ThreadResource {
-    Lock *lock_;
-    LockResource(Lock *lock): ThreadResource() {
-        lock_ = lock;
-    }
-    ~LockResource() {
-        delete lock_;
-    }
-};
-#endif
-
 struct LockObject {
     Lock *lock;
-#ifdef SW_THREAD
-    LockResource *lock_res;
-#endif
     zend_object std;
 };
 
@@ -77,18 +61,9 @@ void php_swoole_lock_set_ptr(zval *zobject, Lock *ptr) {
 
 static void php_swoole_lock_free_object(zend_object *object) {
     LockObject *o = php_swoole_lock_fetch_object(object);
-#ifdef SW_THREAD
-    zend_long resource_id = zend::object_get_long(object, ZEND_STRL("id"));
-    if (o->lock_res && php_swoole_thread_resource_free(resource_id, o->lock_res)) {
-        delete o->lock_res;
-        o->lock_res = nullptr;
-        o->lock = nullptr;
-    }
-#else
     if (o->lock) {
         delete o->lock;
     }
-#endif
     zend_object_std_dtor(object);
 }
 
@@ -110,9 +85,6 @@ static PHP_METHOD(swoole_lock, lock_read);
 static PHP_METHOD(swoole_lock, trylock_read);
 static PHP_METHOD(swoole_lock, unlock);
 static PHP_METHOD(swoole_lock, destroy);
-#ifdef SW_THREAD
-static PHP_METHOD(swoole_lock, __wakeup);
-#endif
 SW_EXTERN_C_END
 
 // clang-format off
@@ -127,20 +99,13 @@ static const zend_function_entry swoole_lock_methods[] =
     PHP_ME(swoole_lock, trylock_read, arginfo_class_Swoole_Lock_trylock_read, ZEND_ACC_PUBLIC)
     PHP_ME(swoole_lock, unlock,       arginfo_class_Swoole_Lock_unlock,       ZEND_ACC_PUBLIC)
     PHP_ME(swoole_lock, destroy,      arginfo_class_Swoole_Lock_destroy,      ZEND_ACC_PUBLIC)
-#ifdef SW_THREAD
-    PHP_ME(swoole_lock, __wakeup,     arginfo_class_Swoole_Lock___wakeup,      ZEND_ACC_PUBLIC)
-#endif
     PHP_FE_END
 };
 // clang-format on
 
 void php_swoole_lock_minit(int module_number) {
     SW_INIT_CLASS_ENTRY(swoole_lock, "Swoole\\Lock", nullptr, swoole_lock_methods);
-#ifndef SW_THREAD
     SW_SET_CLASS_NOT_SERIALIZABLE(swoole_lock);
-#else
-    zend_declare_property_long(swoole_lock_ce, ZEND_STRL("id"), 0, ZEND_ACC_PUBLIC);
-#endif
     SW_SET_CLASS_CLONEABLE(swoole_lock, sw_zend_class_clone_deny);
     SW_SET_CLASS_UNSET_PROPERTY_HANDLER(swoole_lock, sw_zend_class_unset_property_deny);
     SW_SET_CLASS_CUSTOM_OBJECT(
@@ -154,9 +119,6 @@ void php_swoole_lock_minit(int module_number) {
     zend_declare_class_constant_long(swoole_lock_ce, ZEND_STRL("SPINLOCK"), Lock::SPIN_LOCK);
 #endif
     zend_declare_property_long(swoole_lock_ce, ZEND_STRL("errCode"), 0, ZEND_ACC_PUBLIC);
-#ifdef SW_THREAD
-    zend_declare_property_long(swoole_lock_ce, ZEND_STRL("id"), 0, ZEND_ACC_PUBLIC | ZEND_ACC_READONLY);
-#endif
 
     SW_REGISTER_LONG_CONSTANT("SWOOLE_MUTEX", Lock::MUTEX);
 #ifdef HAVE_RWLOCK
@@ -168,47 +130,40 @@ void php_swoole_lock_minit(int module_number) {
 }
 
 static PHP_METHOD(swoole_lock, __construct) {
-    auto o = php_swoole_lock_fetch_object(Z_OBJ_P(ZEND_THIS));
-    if (o->lock != nullptr) {
+    Lock *lock = php_swoole_lock_get_ptr(ZEND_THIS);
+    if (lock != nullptr) {
         zend_throw_error(NULL, "Constructor of %s can only be called once", SW_Z_OBJCE_NAME_VAL_P(ZEND_THIS));
         RETURN_FALSE;
     }
 
-    zend_long type = swoole::Lock::MUTEX;
+    zend_long type = Lock::MUTEX;
+    char *filelock;
+    size_t filelock_len = 0;
 
-    ZEND_PARSE_PARAMETERS_START(0, 1)
-    Z_PARAM_OPTIONAL
-    Z_PARAM_LONG(type)
-    ZEND_PARSE_PARAMETERS_END();
+    if (zend_parse_parameters(ZEND_NUM_ARGS(), "|ls", &type, &filelock, &filelock_len) == FAILURE) {
+        RETURN_FALSE;
+    }
 
-#ifdef SW_THREAD
-    bool process_shared = false;
-#else
-    bool process_shared = true;
-#endif
-
-    Lock *lock;
     switch (type) {
 #ifdef HAVE_SPINLOCK
     case Lock::SPIN_LOCK:
-        lock = new SpinLock(process_shared ? 1 : 0);
+        lock = new SpinLock(1);
         break;
 #endif
 #ifdef HAVE_RWLOCK
     case Lock::RW_LOCK:
-        lock = new RWLock(process_shared ? 1 : 0);
+        lock = new RWLock(1);
         break;
 #endif
     case Lock::MUTEX:
+        lock = new Mutex(Mutex::PROCESS_SHARED);
+        break;
     default:
-        lock = new Mutex(process_shared ? Mutex::PROCESS_SHARED: 0);
+        zend_throw_exception(
+            swoole_exception_ce, "lock type[%d] is not support", type);
+        RETURN_FALSE;
         break;
     }
-#ifdef SW_THREAD
-    o->lock_res = new LockResource(lock);
-    auto resource_id = php_swoole_thread_resource_insert(o->lock_res);
-    zend_update_property_long(swoole_lock_ce, SW_Z8_OBJ_P(ZEND_THIS), ZEND_STRL("id"), resource_id);
-#endif
     php_swoole_lock_set_ptr(ZEND_THIS, lock);
     RETURN_TRUE;
 }
@@ -260,22 +215,7 @@ static PHP_METHOD(swoole_lock, lock_read) {
 }
 
 static PHP_METHOD(swoole_lock, destroy) {
-#ifndef SW_THREAD
     Lock *lock = php_swoole_lock_get_and_check_ptr(ZEND_THIS);
     delete lock;
     php_swoole_lock_set_ptr(ZEND_THIS, nullptr);
-#endif
 }
-
-#ifdef SW_THREAD
-static PHP_METHOD(swoole_lock, __wakeup) {
-    auto o = php_swoole_lock_fetch_object(Z_OBJ_P(ZEND_THIS));
-    zend_long resource_id = zend::object_get_long(ZEND_THIS, ZEND_STRL("id"));
-    o->lock_res = static_cast<LockResource *>(php_swoole_thread_resource_fetch(resource_id));
-    if (!o->lock_res) {
-        zend_throw_exception(swoole_exception_ce, EMSG_NO_RESOURCE, ECODE_NO_RESOURCE);
-        return;
-    }
-    php_swoole_lock_set_ptr(ZEND_THIS, o->lock_res->lock_);
-}
-#endif
