+++ include/swoole_error.h
@@ -36,6 +36,7 @@ enum swErrorCode {
     SW_ERROR_WRONG_OPERATION,
     SW_ERROR_PHP_RUNTIME_NOTICE, // Non-fatal errors, just runtime warnings
     SW_ERROR_UNDEFINED_BEHAVIOR = 600,
+    SW_ERROR_NOT_THREAD_SAFETY,
 
     SW_ERROR_FILE_NOT_EXIST = 700,
     SW_ERROR_FILE_TOO_LARGE,
